# docs/swagger.yaml
definitions:
...
  dto.ReqCheckDuplicatedRole:
    properties:
      excluded_role_info_id:
        type: string
      role_name:
        type: string
    required:
    - role_name
    type: object
...
  dto.ReqCreateRole:
    properties:
      accesses:
        items:
          type: string
        minItems: 1
        type: array
      description:
        type: string
      role_name:
        maxLength: 80
        type: string
    required:
    - accesses
    - role_name
    type: object
...
  dto.ReqUpdateRole:
    properties:
      accesses:
        items:
          type: string
        minItems: 1
        type: array
      description:
        type: string
      role_name:
        maxLength: 80
        type: string
    required:
    - accesses
    - role_name
    type: object
...
  dto.RespPermissionGroup:
    properties:
      id:
        type: string
      name:
        type: string
      value:
        type: boolean
    type: object
  dto.RespPermissionGroupByModule:
    properties:
      accesses:
        items:
          $ref: '#/definitions/dto.RespPermissionGroup'
        type: array
      name:
        $ref: '#/definitions/utils.NullString'
    type: object
...
  dto.RespRole:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
  dto.RespRoleDetail:
    properties:
      created_at:
        type: string
      description:
        $ref: '#/definitions/utils.NullString'
      id:
        type: string
      modules:
        items:
          $ref: '#/definitions/dto.RespPermissionGroupByModule'
        type: array
      role_name:
        type: string
      total_user:
        type: integer
      updated_at:
        $ref: '#/definitions/utils.NullTime'
    type: object
  dto.RespRoleIndex:
    properties:
      created_at:
        type: string
      id:
        type: string
      modules:
        items:
          $ref: '#/definitions/utils.NullString'
        type: array
      role_name:
        type: string
      total_user:
        type: integer
      updated_at:
        $ref: '#/definitions/utils.NullTime'
    type: object
...
  response.NonPaginationResponse:
    properties:
      data: {}
      message:
        type: string
      status:
        type: integer
    type: object
  response.PaginationMeta:
    properties:
      current_page:
        type: integer
      from:
        type: integer
      last_page:
        type: integer
      per_page:
        type: integer
      to:
        type: integer
      total:
        type: integer
    type: object
  response.PaginationResponse:
    properties:
      data: {}
      message:
        type: string
      meta:
        $ref: '#/definitions/response.PaginationMeta'
      status:
        type: integer
    type: object
  utils.NullString:
    properties:
      string:
        type: string
      valid:
        description: Valid is true if String is not NULL
        type: boolean
    type: object
  utils.NullTime:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
info:
  contact: {}
  description: Welcome to the API documentation for the Base Template Web Application.
    This comprehensive guide is designed to help developers seamlessly integrate and
    interact with our platform's functionalities. Whether you're building new features,
    enhancing existing ones, or troubleshooting, this documentation provides all the
    necessary resources and information.
  title: Base Template API Documentation
  version: 0.0-beta
paths:
...
  /v1/role-management/role:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of roles with optional filtering
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: per_page
        type: integer
      - description: Sort column
        in: query
        name: sort_by
        type: string
      - description: Sort order (asc/desc)
        in: query
        name: sort_order
        type: string
      - description: Search query
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved roles
          schema:
            allOf:
            - $ref: '#/definitions/response.PaginationResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.RespRoleIndex'
                  type: array
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
      security:
      - BearerAuth: []
      summary: Get paginated list of roles
      tags:
      - Role Management
    post:
      consumes:
      - application/json
      description: Create a new role with provided information
      parameters:
      - description: Role creation data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ReqCreateRole'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully created role
          schema:
            allOf:
            - $ref: '#/definitions/response.NonPaginationResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.RespRole'
              type: object
        "400":
          description: Bad request - validation error
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
      security:
      - BearerAuth: []
      summary: Create a new role
      tags:
      - Role Management
  /v1/role-management/role/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete a role by its ID
      parameters:
      - description: Role UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully deleted role
          schema:
            allOf:
            - $ref: '#/definitions/response.NonPaginationResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.RespRole'
              type: object
        "400":
          description: Bad request - invalid UUID
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
        "404":
          description: Role not found
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
      security:
      - BearerAuth: []
      summary: Delete role (soft delete)
      tags:
      - Role Management
    get:
      consumes:
      - application/json
      description: Retrieve a specific role by its ID
      parameters:
      - description: Role UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved role
          schema:
            allOf:
            - $ref: '#/definitions/response.NonPaginationResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.RespRoleDetail'
              type: object
        "400":
          description: Bad request - invalid UUID
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
        "404":
          description: Role not found
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
      security:
      - BearerAuth: []
      summary: Get role by ID
      tags:
      - Role Management
    put:
      consumes:
      - application/json
      description: Update an existing role's information
      parameters:
      - description: Role UUID
        in: path
        name: id
        required: true
        type: string
      - description: Updated role data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ReqUpdateRole'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated role
          schema:
            allOf:
            - $ref: '#/definitions/response.NonPaginationResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.RespRole'
              type: object
        "400":
          description: Bad request - validation error
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
        "404":
          description: Role not found
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
      security:
      - BearerAuth: []
      summary: Update role
      tags:
      - Role Management
  /v1/role-management/role/all:
    get:
      consumes:
      - application/json
      description: Retrieve all roles without pagination
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved all roles
          schema:
            allOf:
            - $ref: '#/definitions/response.NonPaginationResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.RespRole'
                  type: array
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
      security:
      - BearerAuth: []
      summary: Get all roles
      tags:
      - Role Management
  /v1/role-management/role/check-name:
    get:
      consumes:
      - application/json
      description: Check if a role name already exists in the database
      parameters:
      - description: Check duplicated role request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ReqCheckDuplicatedRole'
      produces:
      - application/json
      responses:
        "200":
          description: Role with such name exists
          schema:
            allOf:
            - $ref: '#/definitions/response.NonPaginationResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.RespRole'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
        "404":
          description: Role with such name is not found
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
      security:
      - BearerAuth: []
      summary: Check if role name is duplicated
      tags:
      - Role Management
...
securityDefinitions:
  BearerAuth:
    description: 'Enter JWT token (ex: Bearer eyJhbGciOiJIU....)'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"